# [CUDA] Wait for tasks in cuda