This helps keep the work submission from running to far ahead of complete work for the CUDA back-end. I'ts virtually identical to what we do for Metal.

Some benchmarks:

Model | Pre It/sec | Post it/sec | Pre Peak Mem GB | Post Peak Mem GB |
----- | ---- | ---- | ---- | ----
0.6 B  | 6.45 | 6.44 | 54.3 | 28.9
0.86 B  | 3.59 | 5.82 | 64.79 | 36.47

Generation speed is unaffected.